refactor: standardize import paths and rename files for consistency
