+++ drizzle.config.ts
@@ -4,7 +4,7 @@ dotenv.config();
 
 export default defineConfig({
   out: "./src/infrastructure/database/migrations",
-  schema: "./src/infrastructure/database/schema.ts",
+  schema: "./src/infrastructure/database/Schema.ts",
   dialect: "postgresql",
   dbCredentials: {
     url: process.env.DATABASE_URL!,
