+++ src/application/commands/development/level-db-get.ts
@@ -1,5 +1,5 @@
 import { Command } from "@interfaces/discord/Command";
-import { LevelDB } from "@storage/level/client";
+import { LevelDB } from "@storage/level/Client";
 import { ApplicationCommandType, GuildMember } from "discord.js";
 import { container } from "tsyringe";
