+++ src/application/commands/development/level-db-set.ts
@@ -1,5 +1,5 @@
 import { Command } from "@interfaces/discord/Command";
-import { LevelDB } from "@storage/level/client";
+import { LevelDB } from "@storage/level/Client";
 import {
   ApplicationCommandOptionType,
   ApplicationCommandType,
