+++ src/application/events/system/LoadComponents.ts
@@ -1,6 +1,6 @@
 import { container } from "tsyringe";
 import { Event } from "@interfaces/discord/Event";
-import { ClientDiscord } from "@discord/client";
+import { ClientDiscord } from "@discord/Client";
 
 export default new Event({
   name: "interactionCreate",
