+++ src/index.ts
@@ -1,11 +1,11 @@
 import "reflect-metadata";
 import "@container/index";
 import { container } from "tsyringe";
-import { ClientDiscord } from "@discord/client";
-import { Logger } from "@logging/logger";
+import { ClientDiscord } from "@discord/Client";
+import { Logger } from "@logging/Logger";
 import { EventsLoader } from "@discord/loaders/EventsLoader";
 import { CommandLoader } from "@discord/loaders/CommandLoader";
-import { DatabaseConnection } from "@database/connection";
+import { DatabaseConnection } from "@database/Connection";
 
 async function bootstrap() {
   const client = container.resolve(ClientDiscord);
