+++ src/infrastructure/database/Connection.ts
@@ -1,4 +1,4 @@
-import { Logger } from "@logging/logger";
+import { Logger } from "@logging/Logger";
 import { sql } from "drizzle-orm";
 import { drizzle } from "drizzle-orm/node-postgres";
 import { Pool } from "pg";
