+++ src/infrastructure/discord/handlers/CommandHandler.ts
@@ -1,5 +1,5 @@
 import { ApplicationCommandDataResolvable } from "discord.js";
-import { ClientDiscord } from "@discord/client";
+import { ClientDiscord } from "@discord/Client";
 import { injectable, inject } from "tsyringe";
 import { CommandType } from "@interfaces/discord/Command";
