+++ src/infrastructure/storage/level/Client.ts
@@ -1,4 +1,4 @@
-import { Logger } from "@logging/logger";
+import { Logger } from "@logging/Logger";
 import { Level } from "level";
 import { inject, injectable } from "tsyringe";
