+++ src/shared/constants/BasePath.ts
@@ -1,3 +1,3 @@
-import { isDev } from "@utils/isDev";
+import { isDev } from "@utils/IsDev";
 
 export const BASE_PATH = isDev ? "./src" : "./dist";
