+++ src/shared/container/database/index.ts
@@ -1,5 +1,5 @@
 import { container } from "tsyringe";
-import { DatabaseConnection } from "@database/connection";
+import { DatabaseConnection } from "@database/Connection";
 import { ServerEventFlow } from "@database/repositories/ServerEventFlow.repository";
 
 container.registerSingleton(DatabaseConnection, DatabaseConnection);
