+++ src/shared/container/discord/index.ts
@@ -1,11 +1,11 @@
 import { container } from "tsyringe";
-import { Logger } from "@logging/logger";
-import { ClientDiscord } from "@discord/client";
+import { Logger } from "@logging/Logger";
+import { ClientDiscord } from "@discord/Client";
 import { EventHandler } from "@discord/handlers/EventHandler";
 import { EventsLoader } from "@discord/loaders/EventsLoader";
 import { CommandHandler } from "@discord/handlers/CommandHandler";
 import { CommandLoader } from "@discord/loaders/CommandLoader";
-import { BASE_PATH } from "@constants/basePath";
+import { BASE_PATH } from "@constants/BasePath";
 import path from "path";
 
 container.registerSingleton(Logger, Logger);
