+++ src/shared/container/levelDB/index.ts
@@ -1,7 +1,7 @@
 import { container } from "tsyringe";
-import { LevelDB } from "@storage/level/client";
+import { LevelDB } from "@storage/level/Client";
 import path from "path";
-import { BASE_PATH } from "@constants/basePath";
+import { BASE_PATH } from "@constants/BasePath";
 
 container.register("dbPath", {
   useValue: path.resolve(BASE_PATH, "infrastructure/storage/.local"),
